fix: int -> long

+++ src/main/java/ua/com/smiddle/emulator/core/services/processing/Processor.java
@@ -1,11 +1,11 @@
 package ua.com.smiddle.emulator.core.services.processing;
 
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * @author ksa on 09.12.16.
  * @project emulator
  */
 public interface Processor {
-    AtomicInteger getMessageReadCounter();
+    AtomicLong getMessageReadCounter();
 }
